+++ packages/pdf-parser/src/PdfParser.js
@@ -6,7 +6,8 @@ import {
 } from '@ridi/parser-core';
 
 import fs from 'fs';
-import pdfJs from 'pdfjs-dist';
+import pdfJs, { PDFWorker } from 'pdfjs-dist';
+import uuid from 'uuid/v4';
 
 import Book from './model/Book';
 import ParseContext from './model/ParseContext';
@@ -16,14 +17,19 @@ class PdfParser extends Parser {
    * Get default values of parse options
    */
   static get parseDefaultOptions() {
-    return {};
+    return {
+      // Use fake worker when used in a browser environment such as Electron Renderer Proccess.
+      fakeWorker: false,
+    };
   }
 
   /**
    * Get types of parse options
    */
   static get parseOptionTypes() {
-    return {};
+    return {
+      fakeWorker: 'Boolean',
+    };
   }
 
   static get readDefaultOptions() {
@@ -89,6 +95,16 @@ class PdfParser extends Parser {
   }
 
   /**
+   * @returns {ParseTask[]} return after tasks
+   */
+  _parseAfterTasks() {
+    return [
+      { fun: this._destoryWorkerIfNeeded, name: 'destoryWorkerIfNeeded' },
+      ...super._parseAfterTasks(),
+    ];
+  }
+
+  /**
    * @param {object} that
    * @param {function} fun
    * @param {*[]} args
@@ -116,10 +132,12 @@ class PdfParser extends Parser {
    * @throws {Errors.EPDFJS} pdfjs error
    */
   async _loadDocuemnt(context) {
-    const { rawBook, entries } = context;
-    const pdfFile = await entries.first.getFile();
-    const document = await this._execute(pdfJs, pdfJs.getDocument, [pdfFile]);
+    const { rawBook, entries, options } = context;
+    const worker = options.fakeWorker ? new PDFWorker(`pdfWorker_${uuid()}`) : null;
+    const data = await entries.first.getFile();
+    const document = await this._execute(pdfJs, pdfJs.getDocument, [{ data, worker }]);
     context.document = document;
+    context.worker = worker;
     rawBook.pageCount = document.numPages;
     return context;
   }
@@ -196,6 +214,19 @@ class PdfParser extends Parser {
   async _parsePermission(context) {
     const { rawBook, document } = context;
     rawBook.permissions = await this._execute(document, document.getPermissions);
+    return context;
+  }
+
+  /**
+   * Destory fake worker.
+   * @param {ParseContext} context intermediate result
+   * @returns {Promise.<ParseContext>} return Context containing permissions
+   */
+  async _destoryWorkerIfNeeded(context) {
+    const { worker } = context;
+    if (isExists(worker)) {
+      worker.destroy();
+    }
     return context;
   }